+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/BootstrapClient.lua
@@ -1,12 +1,19 @@
 --  #################################################################################################################################
 --  #########                                                STATS CONFIGURATOR                                             #########
---  #################################################################################################################################
-
---  ===================================
+--  =================================================================================================================================
 logSource = "Lua:BootstrapClient"
 Ext.Require("S7_ConfigAuxiliary.lua")
 Ext.Require("S7_StatsConfigurator.lua")
---  ===================================
+--  #################################################################################################################################
+
+local function S7_CatchBroadcast(channel, payload)
+    S7_DebugLog("Client recieved Active Configuration. Saving File.")
+    Ext.SaveFile("S7_ActiveConfig.json", payload)
+end
+
+--  =========================================================
+Ext.RegisterNetListener("S7_ActiveConfig", S7_CatchBroadcast)
+--  =========================================================
 
 local function S7_StatsLoader() --  Loads stats-configuration json during StatsLoaded Event.
     if S7_ConfigSettings.StatsLoader.Enable == true then